Load codemirror CSS via require().
I'm pretty sure our previous comment about @import'ing the CSS
not working (see the deletes in this commit) was because
that @import wasn't actually inlining the CSS; according to
the documentation for postcss-partial-import, which is what precss
was using before we dropped it in #52, @import's are only inlined
if they are of the form @import "foo/bar", and in that case
"foo/_bar.css" will be inlined. Our @import wasn't formatted like that,
so I think that the browser was actually retrieving the codemirror
CSS itself, which caused a race condition.